replace Printer class by veprintf function
