+++ examples/hall/lib/hall-stm32/hall.cpp
@@ -311,6 +311,73 @@ namespace hall {
         auto idx = irqMap[irq-16];
         devMap[idx]->interrupt();
     }
+
+    auto veprintf(void (*fun) (void*,int), void* arg,
+                        char const* fmt, va_list ap) -> int {
+        int pad, count = 0;
+        auto emit = [&](int c) { ++count; fun(arg, c); };
+        auto fill = [&](int n) { while (--n >= 0) emit(pad); };
+
+        while (*fmt)
+            if (char c = *fmt++; c != '%')
+                emit(c);
+            else {
+                pad = *fmt == '0' ? '0' : ' ';
+                int radix = 0;
+                int width = 0;
+                while (radix == 0)
+                    switch (c = *fmt++) {
+                        case 'b': radix =  2; break;
+                        case 'o': radix =  8; break;
+                        case 'u':
+                        case 'd': radix = 10; break;
+                        case 'p': pad = '0'; width = 8;
+                                  [[fallthrough]];
+                        case 'x': radix = 16; break;
+                        case 'c': fill(width - 1);
+                                  c = va_arg(ap, int);
+                                  [[fallthrough]];
+                        case '%': emit(c); radix = 1; break;
+                        case '*': width = va_arg(ap, int); break;
+                        case 's': { char const* s = va_arg(ap, char const*);
+                                    while (*s) {
+                                        emit(*s++);
+                                        --width;
+                                    }
+                                    fill(width);
+                                  }
+                                  [[fallthrough]];
+                        default:  if ('0' <= c && c <= '9')
+                                    width = 10 * width + c - '0';
+                                  else
+                                    radix = 1; // stop scanning
+                    }
+                if (radix > 1) {
+                    static uint8_t numBuf [32]; // shared by all printf's
+                    uint8_t numFill = 0;
+                    int val = va_arg(ap, int);
+                    auto sign = val < 0 && c == 'd';
+                    uint32_t num = sign ? -val : val;
+                    do {
+                        numBuf[numFill++] = "0123456789ABCDEF"[num % radix];
+                        num /= radix;
+                    } while (num != 0);
+                    if (sign) {
+                        if (pad == ' ')
+                            numBuf[numFill++] = '-';
+                        else {
+                            --width;
+                            emit('-');
+                        }
+                    }
+                    fill(width - numFill);
+                    while (numFill > 0)
+                        emit(numBuf[--numFill]);
+                }
+            }
+
+        return count;
+    }
 }
 
 // to re-generate "irqs.h", see the "all-irqs.sh" script
